+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 * Fixed Fashion Loadouts being unable to store an Ornament for the Titan exotic Loreley Splendor Helm.
+* Added button to sort triumphs by completion.
 
 ## 7.75.0 <span class="changelog-date">(2023-07-02)</span>