+++ docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Next
 * Added a toggle button for sorting tabs in the records page.
 
+* Fixed Fashion Loadouts being unable to store an Ornament for the Titan exotic Loreley Splendor Helm.
+
+## 7.75.0 <span class="changelog-date">(2023-07-02)</span>
+
 * Organizer's "Loadouts" column now sorts items by the number of Loadouts using them.
 * Added `memento:none` filter to highlight weapons with an empty memento socket.
 * `deepsight:harmonizable` highlights weapons where Deepsight Resonance can be activated using a Deepsight Harmonizer.